+++ src/data/address.ts
@@ -1,0 +1,15 @@
+// contains types for the resume
+import {Address} from '../types';
+
+export const name = "Andrew Lyle Jones";
+
+export const address: Address = {
+  street: [
+    '2300 24th Rd South',
+    'Apartment 125'
+  ],
+  city: 'Arlington',
+  state: 'Virginia',
+  zipcode: '22206',
+  country: 'U.S.A'
+};
