+++ src/data/frameworks.ts
@@ -1,0 +1,30 @@
+import { Framework } from '../types';
+
+export const vue: Framework = {
+  name: 'Vue.JS',
+  experience: 5
+}
+
+export const react: Framework = {
+  name: 'React.JS',
+  experience: 2
+}
+
+export const angular: Framework = {
+  name: 'Angular.JS',
+  experience: 3
+}
+
+export const capacitor: Framework = {
+  name: 'Capacitor.JS',
+  experience: 2
+}
+
+export const wordpress: Framework = {
+  name: 'Wordpress',
+  experience: 2
+}
+
+export const frameworks: Array<Framework> = {
+  vue, react, angular, capacitor, wordpress
+}
